2022-08-13 - Large Amount of Changes
Okay I did a huge amount of changes for this version. The general gist of it is that previously the dodman generator was a single script that the user runs. That was not really good for the following reasons:
-Not very portable - you could not really import the script to be used in other python scripts
-Very difficult to resume generation after you start

What I did this time is move the whole script into a single class called 'DodmanGen'. The points, paths and 'coin/dice' flip history are all stored in the single class allowing not just the generator but generated map to be easily saved or stored by the user. This is also for future plans to add the generator onto a user interface.

Aside from that I also made the following value name changes for ease of reading and understanding:
startcoord > start_coord
randcoord > random_start
maptype > map_shape
size > point_rad
zeropow > zero_pow
flipres > flip_results
maxflip > max_label_length

Have fun